Added merging behavior for `dest`

+++ directory_sync.py
@@ -1,9 +1,10 @@
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
 Created on Sat Oct 21 09:24:35 2023
 
 @author: tjdwill
-@revision: 12 March 2024
+@revision: 20 December 2024
 @description: Command-line tool to automate synchronization of two
 same-named directories stored in different locations.
 """
@@ -40,6 +41,7 @@ basic_fmt = logging.Formatter(fmt="%(asctime)s\n%(message)s", datefmt='%d-%m-%Y 
 file_handle.setFormatter(basic_fmt)
 log.addHandler(file_handle)
 
+
 # Argument Parsing Config
 parser = argparse.ArgumentParser(
     # prog="DirectorySync",
@@ -60,9 +62,9 @@ parser.add_argument(
     )
 )
 parser.add_argument(
-    '-y', '--skip-confirmation',
-    default=False,
-    help="Run without FROM-TO directional confirmation.",
+    '-m', '--merge',
+    "Keeps files that are unique to `dest` instead of deleting them.",
+    default=False, 
     action="store_true"
 )
 parser.add_argument(
@@ -73,6 +75,12 @@ parser.add_argument(
         *["DEBUG", "INFO", "WARN", "ERROR", "CRIT"]
     ],
     default="WARN",
+)
+parser.add_argument(
+    '-y', '--skip-confirmation',
+    default=False,
+    help="Run without FROM-TO directional confirmation.",
+    action="store_true"
 )
 
 
@@ -94,7 +102,8 @@ Logic:
 def dir_comp(
         src: Path,
         dest: Path,
-        curr_list: list
+        curr_list: list,
+        merge: bool
 ) -> list:
     """
     Outputs an updated update_list
@@ -108,11 +117,15 @@ def dir_comp(
     Parameters
     ----------
     src : Path
+        The directory to sync with. At the end of a successful run, `dest` should match `src`.
         Assumes an absolute path.
     dest : Path
         Destination path. Assumes absolute path.
     curr_list : list
         List of directories to traverse.
+    merge : bool
+        Triggers merging behavior. When True, the program keeps files that are 
+        in *dest* but not in *src* (instead of removing them).
 
     Returns
     -------
@@ -135,19 +148,20 @@ def dir_comp(
     shared_items = src_items.intersection(dest_items)
 
     # Remove from dest all items not in src
-    for item in dest_unique:
-        path = dest / item
-        try:
-            if path.is_dir():
-                shutil.rmtree(path)
+    if not merge:
+        for item in dest_unique:
+            path = dest / item
+            try:
+                if path.is_dir():
+                    shutil.rmtree(path)
+                else:
+                    # Remove the file
+                    path.unlink()
+            except PermissionError as e:
+                log.warning(f"Could not delete '{path}'.\n{e}")
+                error_found = True
             else:
-                # Remove the file
-                path.unlink()
-        except PermissionError as e:
-            log.warning(f"Could not delete '{path}'.\n{e}")
-            error_found = True
-        else:
-            log.info(f"Deleted '{path}'.")
+                log.info(f"Deleted '{path}'.")
 
     # Copy unique items to dest
     for item in src_unique:
@@ -201,14 +215,15 @@ if __name__ == '__main__':
 
     # Parse Command-line
     args = parser.parse_args()
-    source_path, dest_path = args.source, args.dest
-    log_lvl = args.log_level.upper()
-    skip_confirmation = args.skip_confirmation
+    SOURCE_PATH, DEST_PATH = args.source, args.dest
+    LOG_LVL = args.log_level.upper()
+    SKIP_CONFIRMATION = args.skip_confirmation
+    MERGE = args.merge
 
 
     # Initial Checks
-    src = Path(source_path).resolve()
-    dest = Path(dest_path).resolve()
+    src = Path(SOURCE_PATH).resolve()
+    dest = Path(DEST_PATH).resolve()
     if not src.is_dir():
         print(f"[ERROR] Source folder: \"{src}\" is not a valid directory.\n")
         raise OSError
@@ -227,9 +242,12 @@ if __name__ == '__main__':
 
     # Ensure desired directionality
     confirmed = False
-    while (not confirmed) and (not skip_confirmation):
+    while (not confirmed) and (not SKIP_CONFIRMATION):
         print(f'Folders to sync:\n  "FROM: {src}"\n  "TO: {dest}"\n')
-        user_input = input("Files in destination folder may be deleted. Are you sure this the correct sync direction?  (y/n): ").lower()
+        user_input = input(
+            "Files in destination folder may be deleted. "
+            "Are you sure this the correct sync direction?  (y/n): "
+        ).lower()
         if user_input == 'y':
             confirmed = True
             continue
@@ -243,16 +261,21 @@ if __name__ == '__main__':
     # Begin Program
     log.info(f"Directory Sync\nFROM: {src}\nTO: {dest}\nBegin.")
     file_handle.setFormatter(main_fmtr)
-    file_handle.setLevel(log_levels[log_lvl])
+    file_handle.setLevel(log_levels[LOG_LVL])
 
-    INFORM_USER = False  # Did an error get written to the log file?
+    INFORM_USER = False  # Was an error written to the log file?
     update_list = [src]
     iter_src = src
     iter_dest = dest
     done = False
     while not done:
         print(f'Traversing: "{iter_src}"')
-        update_list, error_found = dir_comp(iter_src, iter_dest, update_list)
+        update_list, error_found = dir_comp(
+            iter_src,
+            iter_dest,
+            update_list,
+            MERGE,
+        )
         if error_found:
             INFORM_USER = True
         if not update_list:
